* 'dev_marc' of gh:Catrobat/Catty:
  LoadingCell-Indicator started
  initial commit of segmented control at DownloadController
  changes concerning iOS7.1 update
  better Fullscreen effect at forumWebViewController

Conflicts:
	src/Catty/ForumWebViewController.m
	src/Catty/ProgramTableViewController.m
	src/Catty/RecentProgramsStoreViewController.m
	src/Catty/iPhone.storyboard

+++ src/Catty/AudioManager.m
@@ -102,8 +102,8 @@ static AudioManager* sharedAudioManager = nil;
           return;
 
       }
-      [player setKey:[fileName stringByAppendingString:[NSString stringWithFormat:@"%d",self.soundCounter]]];
-      [audioPlayers setObject:player forKey:[fileName stringByAppendingString:[NSString stringWithFormat:@"%d",self.soundCounter]]];
+      [player setKey:[fileName stringByAppendingString:[NSString stringWithFormat:@"%ld",(long)self.soundCounter]]];
+      [audioPlayers setObject:player forKey:[fileName stringByAppendingString:[NSString stringWithFormat:@"%ld",(long)self.soundCounter]]];
   }
 //  if ([player isPlaying]) {
 //    [player stop];
