+++ src/Catty/Scene.m
@@ -121,7 +121,7 @@
         return NO;
     }
     SpriteObject *obj1 = nodesAtPoint[[nodesAtPoint count]-1];
-    int counter = [nodesAtPoint count]-2;
+    NSInteger counter =[nodesAtPoint count]-2;
     NSDebug(@"How many nodes are touched: %d",counter);
     NSDebug(@"First Node:%@",obj1);
     if (!obj1.name) {
